feat: Refactor Plotly imports and data handling
Replaces dynamic imports of `react-plotly.js` with a dedicated `PlotComponent` for cleaner code and uses `plotly.js-dist-min` to avoid fetch-related SSR issues.

Additionally, updates data processing in `data/features.py` to correctly handle potential NA values during type casting for feature extraction, ensuring robustness.

+++ data/features.py
@@ -570,7 +570,7 @@ class FeaturePipeline:
         missing = [c for c in self.OBS_COLUMNS if c not in df.columns]
         if missing:
             raise ValueError(f"Missing feature columns: {missing}")
-        X = df[self.OBS_COLUMNS].values.astype(np.float32)
+        X = df[self.OBS_COLUMNS].astype('float64').values.astype(np.float32)
         X = np.nan_to_num(X, nan=0.0, posinf=1.0, neginf=-1.0)
         return X
